# docs/openapi-mvp.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/exceptions/{id}": {
      "get": {
        "operationId": "ExceptionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Exception UUID",
            "schema": {
              "example": "e1111111-1111-1111-1111-111111111111",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get exception details and review history",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/exceptions/{id}/review": {
      "post": {
        "operationId": "ExceptionsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Exception UUID",
            "schema": {
              "example": "e1111111-1111-1111-1111-111111111111",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions (Operator cannot review)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit QA review action, notes, and corrective action (Restricted to QA_REVIEWER & ADMIN)",
        "tags": [
          "exceptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "ColdProof API • Cold Chain Evidence Platform",
    "description": "APEX • GenD Arena 2026 • Technical Validation Specification v3.2.\n\nEvidence workflow for pharma cold chain:\n- Ingestion & adapters (Zenodo, Mendeley, Vendor-inspired formats)\n- Canonical normalization with immutable provenance\n- Batch-centric timeline & segment mapping (CP-DEMO-001)\n- Exception engine (DEMO_2_8C product profile)\n- QA human-in-the-loop review actions\n- Evidence package reports with SHA-256 integrity hash\n- Immutable audit trail",
    "version": "3.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReviewActionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "REVIEWED",
              "NEEDS_EVIDENCE",
              "REJECTED"
            ],
            "example": "REVIEWED"
          },
          "reviewer_id": {
            "type": "string",
            "example": "00000000-0000-0000-0000-000000000003"
          },
          "notes": {
            "type": "string",
            "example": "Reviewed temperature excursion during handover."
          },
          "corrective_action": {
            "type": "string",
            "example": "Quarantine batch and verify secondary logger."
          }
        },
        "required": [
          "status"
        ]
      },
      "ReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "exception_id": {
            "type": "string"
          },
          "reviewer_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "exception_id",
          "reviewer_id",
          "status",
          "created_at"
        ]
      },
...
    }
  }
}